requirements.txt
streamlit==1.28.2
packaging==23.2
wheel==0.41.3
typing-extensions==4.10.0

torch==2.2.2
torchvision==0.17.2
transformers==4.36.2
tokenizers==0.15.0
sentencepiece==0.1.99
protobuf==4.25.1

spacy==3.8.11
pydantic==2.7.4

fer==25.10.3
opencv-python==4.9.0.80
facenet-pytorch==2.6.0

sumy==0.11.0
deep-translator==1.11.4
pdfplumber
python-docx
langdetect==1.0.9
SpeechRecognition==3.10.4
pyttsx3==2.90
textblob==0.18.0.post0
pyaudio
